+++ app/index.js
@@ -16,6 +16,7 @@ import { initializePushNotifications, onNotification } from './notifications/pus
 import store from './lib/createStore';
 import NotificationBadge from './notifications/inApp';
 import { defaultHeader, onNavigationStateChange } from './utils/navigation';
+import Toast from './helperComponent/Toast';
 
 useScreens();
 
@@ -231,6 +232,7 @@ class CustomInsideStack extends React.Component {
 			<React.Fragment>
 				<InsideStackModal navigation={navigation} />
 				<NotificationBadge navigation={navigation} />
+				<Toast />
 			</React.Fragment>
 		);
 	}
