+++ app/views/RoomInfoEditView/index.js
@@ -12,7 +12,9 @@ import KeyboardView from '../../presentation/KeyboardView';
 import sharedStyles from '../Styles';
 import styles from './styles';
 import scrollPersistTaps from '../../utils/scrollPersistTaps';
-import { showErrorAlert, Toast } from '../../utils/info';
+import { showErrorAlert } from '../../utils/info';
+import { LISTENER } from '../../helperComponent/Toast';
+import EventEmitter from '../../utils/events';
 import database, { safeAddListener } from '../../lib/realm';
 import RocketChat from '../../lib/rocketchat';
 import RCTextInput from '../../containers/TextInput';
@@ -215,7 +217,7 @@ export default class RoomInfoEditView extends React.Component {
 			if (error) {
 				showErrorAlert(I18n.t('There_was_an_error_while_action', { action: I18n.t('saving_settings') }));
 			} else {
-				this.toast.show(I18n.t('Settings_succesfully_changed'));
+				EventEmitter.emit(LISTENER, { message: I18n.t('Settings_succesfully_changed') });
 			}
 		}, 100);
 	}
@@ -428,7 +430,6 @@ export default class RoomInfoEditView extends React.Component {
 							<Text style={[sharedStyles.button_inverted, styles.colorDanger]} accessibilityTraits='button'>{I18n.t('DELETE')}</Text>
 						</TouchableOpacity>
 						<Loading visible={saving} />
-						<Toast ref={toast => this.toast = toast} />
 					</SafeAreaView>
 				</ScrollView>
 			</KeyboardView>
